[arcane,core] Move 'ItemGroupSubPartsByType' in its own file.
