+++ arcane/src/arcane/core/ItemGroupImpl.cc
@@ -1394,43 +1394,6 @@ hasComputeFunctor() const
 /*---------------------------------------------------------------------------*/
 /*---------------------------------------------------------------------------*/
 
-void ItemGroupSubPartsByType::
-_initChildrenByTypeV1()
-{
-  ItemGroupImpl* p = m_group_impl;
-  IItemFamily* family = m_group_internal->m_item_family;
-  ItemTypeMng* type_mng = family->mesh()->itemTypeMng();
-  Integer nb_basic_item_type= ItemTypeMng::nbBasicItemType(); //NB_BASIC_ITEM_TYPE
-  m_children_by_type.resize(nb_basic_item_type);
-  for (Integer i = 0; i < nb_basic_item_type; ++i) {
-    // String child_name(name()+i);
-    String child_name(type_mng->typeName(i));
-    auto* functor = new ItemGroupImplItemGroupComputeFunctor(p, &ItemGroupImpl::_computeChildrenByType);
-    ItemGroupImpl* igi = m_group_impl->createSubGroup(child_name, family, functor);
-    m_children_by_type[i] = igi;
-  }
-}
-
-/*---------------------------------------------------------------------------*/
-/*---------------------------------------------------------------------------*/
-
-void ItemGroupSubPartsByType::
-_initChildrenByTypeV2()
-{
-  bool is_verbose = m_is_debug_apply_operation;
-  if (is_verbose)
-    m_group_internal->mesh()->traceMng()->info() << "ItemGroupImpl::_initChildrenByTypeV2() name=" << m_group_internal->name();
-
-  Int32 nb_basic_item_type= ItemTypeMng::nbBasicItemType();
-  m_children_by_type_ids.resize(nb_basic_item_type);
-  for (Integer i = 0; i < nb_basic_item_type; ++i) {
-    m_children_by_type_ids[i] = UniqueArray<Int32>{ MemoryUtils::getDefaultDataAllocator() };
-  }
-}
-
-/*---------------------------------------------------------------------------*/
-/*---------------------------------------------------------------------------*/
-
 void ItemGroupImpl::
 _computeChildrenByType()
 {
@@ -1472,74 +1435,6 @@ _computeChildrenByTypeV1()
 /*---------------------------------------------------------------------------*/
 /*---------------------------------------------------------------------------*/
 
-void ItemGroupSubPartsByType::
-_computeChildrenByTypeV2()
-{
-  ItemGroup that_group(m_group_impl);
-  Int32 nb_item = m_group_internal->nbItem();
-  IMesh* mesh = m_group_internal->mesh();
-  ItemTypeMng* type_mng = mesh->itemTypeMng();
-  ITraceMng* trace = mesh->traceMng();
-  bool is_verbose = m_is_debug_apply_operation;
-  if (is_verbose)
-    trace->info() << "ItemGroupImpl::_computeChildrenByTypeV2 for " << m_group_internal->name();
-
-  // Si le maillage est cartésien, on sait que toutes les entités ont le même type
-  if (nb_item > 0 && mesh->meshKind().meshStructure() == eMeshStructure::Cartesian) {
-    ItemInfoListView lv(m_group_internal->m_item_family->itemInfoListView());
-    m_unique_children_type = ItemTypeId{ lv.typeId(m_group_internal->itemsLocalId()[0]) };
-    return;
-  }
-
-  Int32 nb_basic_item_type = ItemTypeMng::nbBasicItemType();
-  m_unique_children_type = ItemTypeId{ IT_NullType };
-
-  UniqueArray<Int32> nb_items_by_type(nb_basic_item_type);
-  nb_items_by_type.fill(0);
-  ENUMERATE_(Item,iitem,that_group){
-    Item item = *iitem;
-    Int16 item_type = item.type();
-    if (item_type<nb_basic_item_type)
-      ++nb_items_by_type[item_type];
-  }
-const String& name = m_group_internal->name();
-  Int32 nb_different_type = 0;
-for (Int32 i = 0; i < nb_basic_item_type; ++i) {
-    m_children_by_type_ids[i].clear();
-    const Int32 n = nb_items_by_type[i];
-    m_children_by_type_ids[i].reserve(n);
-    if (n > 0)
-      ++nb_different_type;
-    if (is_verbose)
-      trace->info() << "ItemGroupImpl::_computeChildrenByTypeV2 for " << name
-                    << " type=" << type_mng->typeName(i) << " nb=" << n;
-  }
-  if (is_verbose)
-    trace->info() << "ItemGroupImpl::_computeChildrenByTypeV2 for " << name
-                  << " nb_item=" << nb_item << " nb_different_type=" << nb_different_type;
-
-  // Si nb_different_type == 1, cela signifie qu'il n'y a qu'un seul
-  // type d'entité et on conserve juste ce type, car dans ce cas on passera
-  // directement le groupe en argument de applyOperation().
-  if (nb_item > 0 && nb_different_type == 1) {
-    ItemInfoListView lv(m_group_internal->m_item_family->itemInfoListView());
-    m_unique_children_type = ItemTypeId{ lv.typeId(m_group_internal->itemsLocalId()[0]) };
-    if (is_verbose)
-      trace->info() << "ItemGroupImpl::_computeChildrenByTypeV2 for " << name
-                    << " unique_type=" << type_mng->typeName(m_unique_children_type);
-    return;
-  }
-
-  ENUMERATE_(Item,iitem,that_group){
-    Item item = *iitem;
-    Integer item_type = item.type();
-    if (item_type<nb_basic_item_type)
-      m_children_by_type_ids[item_type].add(iitem.itemLocalId());
-  }
-
-  for( Int32 i=0; i<nb_basic_item_type; ++i )
-    applySimdPadding(m_children_by_type_ids[i]);
-}
 
 /*---------------------------------------------------------------------------*/
 /*---------------------------------------------------------------------------*/
@@ -1819,6 +1714,26 @@ ItemGroupImplInternal* ItemGroupImpl::
 _internalApi() const
 {
   return &m_p->m_internal_api;
+}
+
+
+/*---------------------------------------------------------------------------*/
+/*---------------------------------------------------------------------------*/
+
+//TODO a supprimer lorsque la V1 ne sera plus disponible
+void ItemGroupSubPartsByType::
+_initChildrenByTypeV1()
+{
+  IItemFamily* family = m_group_internal->m_item_family;
+  ItemTypeMng* type_mng = family->mesh()->itemTypeMng();
+  Integer nb_basic_item_type = ItemTypeMng::nbBasicItemType(); //NB_BASIC_ITEM_TYPE
+  m_children_by_type.resize(nb_basic_item_type);
+  for (Integer i = 0; i < nb_basic_item_type; ++i) {
+    String child_name(type_mng->typeName(i));
+    auto* functor = new ItemGroupImplItemGroupComputeFunctor(m_group_impl, &ItemGroupImpl::_computeChildrenByType);
+    ItemGroupImpl* igi = m_group_impl->createSubGroup(child_name, family, functor);
+    m_children_by_type[i] = igi;
+  }
 }
 
 /*---------------------------------------------------------------------------*/
